refactor: consolidate MLPEncoder, MLPDecoder and DenseModel into a single MLP class
