+++ src/spherinator/models/__init__.py
@@ -11,12 +11,10 @@ from .convolutional_decoder_1d import ConvolutionalDecoder1D
 from .convolutional_decoder_2d import ConvolutionalDecoder2D
 from .convolutional_encoder_1d import ConvolutionalEncoder1D
 from .convolutional_encoder_2d import ConvolutionalEncoder2D
-from .dense_model import DenseModel
 from .embedding_reconstruction import EmbeddingReconstruction
 from .export_onnx import export_onnx
 from .huggingface_vit_encoder import HuggingFaceViTEncoder
-from .mlp_decoder import MLPDecoder
-from .mlp_encoder import MLPEncoder
+from .mlp import MLP
 from .upsampling_decoder_2d import UpsamplingDecoder2D
 from .variational_autoencoder import VariationalAutoencoder
 from .weights_provider import WeightsProvider
@@ -32,9 +30,7 @@ __all__ = [
     "ConvolutionalDecoder2D",
     "ConvolutionalEncoder1D",
     "ConvolutionalEncoder2D",
-    "DenseModel",
-    "MLPDecoder",
-    "MLPEncoder",
+    "MLP",
     "EmbeddingReconstruction",
     "HuggingFaceViTEncoder",
     "UpsamplingDecoder2D",
